inline all of tests/cfy.py
now that we've moved .invoke into the base class, there's only a few
methods left in there, so we can similarly inline all of them, and
do away with the concept of cfy.py entirely
